master : changed table columns and migration. fix for flyway config, changed id type int to long in enities

+++ business-module/db/migration/V1_0__create_user_task_usertask_schemas.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `user`
 (
-    `id`        INT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id`        BIGINT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `userName`  VARCHAR(20) NOT NULL UNIQUE,
     `firstName` VARCHAR(30),
     `lastName`  VARCHAR(30)
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS `user`
 
 CREATE TABLE IF NOT EXISTS `task`
 (
-    `id`          INT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    `id`          BIGINT         NOT NULL AUTO_INCREMENT PRIMARY KEY,
     `title`       VARCHAR(20) NOT NULL,
     `description` VARCHAR(30)
 ) ENGINE = InnoDB
@@ -17,8 +17,8 @@ CREATE TABLE IF NOT EXISTS `task`
 
 CREATE TABLE IF NOT EXISTS `user_task`
 (
-    `user_id` INT NOT NULL,
-    `task_id` INT NOT NULL,
+    `user_id` BIGINT NOT NULL,
+    `task_id` BIGINT NOT NULL,
     FOREIGN KEY (user_id) REFERENCES user (id),
     FOREIGN KEY (task_id) REFERENCES task (id)
 ) ENGINE = InnoDB
